+++ tools/Equinox.Tool/Program.fs
@@ -217,7 +217,8 @@ module LoadTest =
                 decorateWithLogger (log,verbose) execForClient
             | Some storeConfig, _ ->
                 let services = ServiceCollection()
-                Samples.Infrastructure.Services.register(services, storeConfig, storeLog)
+                let codecGen : Services.ICodecGen = Services.NewtonsoftJsonCodecGen() :> _ 
+                Samples.Infrastructure.Services.register(services, storeConfig, storeLog, codecGen)
                 let container = services.BuildServiceProvider()
                 let execForClient = Tests.executeLocal container test
                 decorateWithLogger (log, verbose) execForClient